+++ lldb/tools/lldb-dap/Handler/GotoRequestHandler.cpp
@@ -19,6 +19,9 @@
 #include "llvm/Support/Error.h"
 #include <optional>
 
+using namespace lldb;
+using namespace lldb_dap::protocol;
+
 namespace lldb_dap {
 
 /// The request sets the location where the debuggee will continue to run.
@@ -31,35 +34,32 @@ namespace lldb_dap {
 /// Clients should only call this request if the corresponding capability
 /// `supportsGotoTargetsRequest` is true (because only then goto targets exist
 /// that can be passed as arguments).
-llvm::Error GotoRequestHandler::Run(const protocol::GotoArguments &args) const {
+llvm::Error GotoRequestHandler::Run(const GotoArguments &args) const {
   if (dap.ProcessIsNotStopped())
     return llvm::make_error<NotStoppedError>();
 
-  lldb::SBThread thread = dap.GetLLDBThread(args.threadId);
+  SBThread thread = dap.GetLLDBThread(args.threadId);
   if (!thread.IsValid())
     return llvm::make_error<DAPError>(
         llvm::formatv("invalid thread {}", args.threadId));
 
-  std::optional<lldb::SBLineEntry> entry_opt =
+  std::optional<SBLineEntry> entry_opt =
       dap.goto_targets.GetLineEntry(args.targetId);
   if (!entry_opt)
     return llvm::make_error<DAPError>(
         llvm::formatv("invalid goto target {}", args.targetId));
 
-  const lldb::SBLineEntry &entry = entry_opt.value();
-  lldb::SBFileSpec file_spec = entry.GetFileSpec();
-  uint32_t line = entry.GetLine();
-  // TODO: pass column too, once SBThread::JumpToLine (or a successor)
-  // accepts a column argument.
-
-  lldb::SBError error = thread.JumpToLine(file_spec, line);
+  const SBLineEntry &entry = entry_opt.value();
+  SBFileSpec file_spec = entry.GetFileSpec();
+  const uint32_t line = entry.GetLine();
+  SBError error = thread.JumpToLine(file_spec, line);
   if (error.Fail())
     return ToError(error);
 
-  protocol::StoppedEventBody body;
+  StoppedEventBody body{};
   body.reason = protocol::eStoppedReasonGoto;
   body.threadId = thread.GetThreadID();
-  dap.Send(protocol::Event{"stopped", std::move(body)});
+  dap.Send(Event{"stopped", std::move(body)});
 
   return llvm::Error::success();
 }