+++ lldb/tools/lldb-dap/Protocol/ProtocolRequests.cpp
@@ -474,14 +474,14 @@ llvm::json::Value toJSON(const StepInTargetsResponseBody &SITR) {
   return llvm::json::Object{{"targets", SITR.targets}};
 }
 
-bool fromJSON(const llvm::json::Value &Params, GotoTargetsArguments &GTA,
-              llvm::json::Path P) {
+bool fromJSON(const json::Value &Params, GotoTargetsArguments &GTA,
+              json::Path P) {
   json::ObjectMapper OM(Params, P);
   return OM && OM.map("source", GTA.source) && OM.map("line", GTA.line) &&
          OM.mapOptional("column", GTA.column);
 }
 
-llvm::json::Value toJSON(const GotoTargetsResponseBody &GTR) {
+json::Value toJSON(const GotoTargetsResponseBody &GTR) {
   return llvm::json::Object{{"targets", GTR.targets}};
 }